requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.7
django-cors-headers==3.14.0
django-remote-forms==0.0.1
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
netaddr==0.8.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.15.0
yookassa==2.3.5